+++ flutter_local_notifications/example/lib/main.dart
@@ -256,7 +256,6 @@ class _HomePageState extends State<HomePage> {
       TextEditingController();
 
   bool _notificationsEnabled = false;
-  bool _exactAlarmsEnabled = false;
 
   @override
   void initState() {
@@ -308,13 +307,6 @@ class _HomePageState extends State<HomePage> {
           await androidImplementation?.requestPermission();
       setState(() {
         _notificationsEnabled = grantedNotificationPermission ?? false;
-      });
-
-      final bool? grantedExactAlarmsPermission =
-          await androidImplementation?.requestExactAlarmsPermission();
-
-      setState(() {
-        _exactAlarmsEnabled = grantedExactAlarmsPermission ?? false;
       });
     }
   }
@@ -576,7 +568,6 @@ class _HomePageState extends State<HomePage> {
                       style: TextStyle(fontWeight: FontWeight.bold),
                     ),
                     Text('notifications enabled: $_notificationsEnabled'),
-                    Text('exact alarms enabled: $_exactAlarmsEnabled'),
                     PaddedElevatedButton(
                       buttonText:
                           'Check if notifications are enabled for this app',